+++ src/main/java/browsersupport/Parallel/EdgeSetUp.java
@@ -1,16 +1,16 @@
-package utils.Parallel;
+package browsersupport.Parallel;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.edge.EdgeOptions;
 import org.openqa.selenium.remote.RemoteWebDriver;
-import utils.proxy.ProxyConfig;
+import browsersupport.proxy.ProxyConfig;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 
-import static utils.Browser.*;
+import static browsersupport.Browser.*;
 
 
 public class EdgeSetUp {
