+++ src/main/java/browsersupport/Parallel/FirefoxSetUp.java
@@ -1,16 +1,16 @@
-package utils.Parallel;
+package browsersupport.Parallel;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.firefox.FirefoxOptions;
 import org.openqa.selenium.remote.RemoteWebDriver;
-import utils.proxy.ProxyConfig;
+import browsersupport.proxy.ProxyConfig;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 
-import static utils.Browser.*;
+import static browsersupport.Browser.*;
 
 public class FirefoxSetUp {
