+++ src/main/java/browsersupport/Parallel/SafariSetUp.java
@@ -1,4 +1,4 @@
-package utils.Parallel;
+package browsersupport.Parallel;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
@@ -9,7 +9,7 @@ import org.openqa.selenium.safari.SafariOptions;
 import java.net.MalformedURLException;
 import java.net.URL;
 
-import static utils.Browser.*;
+import static browsersupport.Browser.*;
 
 public class SafariSetUp {
