+++ src/main/java/browsersupport/config/Configuration.java
@@ -1,4 +1,4 @@
-package utils.config;
+package browsersupport.config;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
